Move POST playlist functionality to Create Playlist Modal
