+++ src/components/CreatePlaylistModal/CreatePlaylistModal.tsx
@@ -1,12 +1,94 @@
+import { useState } from "react";
+import axios from "axios";
+
 import "./CreatePlaylistModal.scss";
 import Button from "../Button/Button";
 import closeIcon from "../../assets/close.svg";
 
+const USER_ID_ENDPOINT = "https://api.spotify.com/v1/me";
+
 export default function CreatePlaylistModal({
-  playlistInfo,
-  handlePlaylistInfoChange,
+  accessToken,
   closeModal,
+  tracksToDisplay,
+  playlistInfo,
+  setPlaylistInfo,
 }) {
+  // Function to get userId
+  const getUserId = async () => {
+    const userId = await axios
+      .get(USER_ID_ENDPOINT, {
+        headers: {
+          Authorization: `Bearer ${accessToken}`,
+        },
+      })
+      .then((response) => {
+        console.log(response.data.id);
+        return response.data.id;
+      })
+      .catch((error) => {
+        console.log(error.response.data.error);
+      });
+    return userId;
+  };
+
+  // Function to handle Playlist creating info form changes
+  const handlePlaylistInfoChange = (event) => {
+    let value = event.target.value;
+    if (event.target.name === "public") {
+      if (event.target.value === "private") {
+        value = false;
+      } else {
+        value = true;
+      }
+    }
+    setPlaylistInfo({
+      ...playlistInfo,
+      [event.target.name]: value,
+    });
+  };
+
+  // Function to POST new filtered playlist
+  const handlePostPlaylist = async (event) => {
+    event.preventDefault();
+    const userId = await getUserId();
+    axios
+      .post(
+        `https://api.spotify.com/v1/users/${userId}/playlists`,
+        playlistInfo,
+        {
+          headers: {
+            Authorization: `Bearer ${accessToken}`,
+          },
+        }
+      )
+      .then((response) => {
+        const playlistId = response.data.id;
+        const trackIds = tracksToDisplay.map((track) => {
+          return track.uri;
+        });
+        axios
+          .post(
+            `https://api.spotify.com/v1/playlists/${playlistId}/tracks`,
+            {
+              uris: trackIds,
+              position: 0,
+            },
+            {
+              headers: {
+                Authorization: `Bearer ${accessToken}`,
+              },
+            }
+          )
+          .then((response) => {
+            closeModal();
+          });
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  };
+
   return (
     <div className="create-playlist-modal">
       <form className="create-playlist-modal__form">
@@ -44,6 +126,8 @@ export default function CreatePlaylistModal({
               name="public"
               id="private"
               value="private"
+              onChange={handlePlaylistInfoChange}
+              checked={!playlistInfo.public}
             />
             <label
               className="create-playlist-modal__label-radio"
@@ -60,6 +144,8 @@ export default function CreatePlaylistModal({
               name="public"
               id="public"
               value="public"
+              onChange={handlePlaylistInfoChange}
+              checked={playlistInfo.public}
             />
             <label
               className="create-playlist-modal__label-radio"
@@ -69,7 +155,7 @@ export default function CreatePlaylistModal({
             </label>
           </div>
         </div>
-        <Button variant="primary" text="CREATE" />
+        <Button variant="primary" text="CREATE" onClick={handlePostPlaylist} />
       </form>
     </div>
   );
