+++ src/components/SliderMarks/SliderMarks.scss
@@ -19,8 +19,8 @@
     align-items: center;
     margin-top: 0rem;
     width: 2.5rem;
-    height: 2.5rem;
-    border-radius: 50%;
+    height: 1.5rem;
+    border-radius: 4px;
     transition: all 0.2s ease-in-out;
 
     &--selected {
